performance: speedup computation of suspended revisions
In their current state, revset calls can be very costly, as we test
predicates on the entire repository.

This change drops a revset call in favor of direct testing of the
phase of changesets.

Performance test on my Mercurial checkout
- 19857 total changesets,
- 1584 obsolete changesets,
- 13310 obsolescence markers.

Before:
  ! suspended
  ! wall 0.014319

After:
  ! suspended
  ! wall 0.009559

Performance test on a Mozilla central checkout:
- 117293 total changesets,
- 1 obsolete changeset,
- 1 obsolescence marker.

Before:
  ! suspended
  ! wall 0.033373

After:
  ! suspended
  ! wall 0.000053