bundlerepo: remove old style check of file names
The old check is no longer needed:

The checked paths will be passed on to filelog.__init__ which will prefix the
path with 'data/'. A leading '/' will thus not do any harm.

And: The paths will be used by an opener which will use a pathauditor.

And finally: The old check did not consider Windows paths and was thus
insufficient.