run-tests: fixes the number of tests ran when '--retest' is enabled
This patch fixes a regression recently introduced by a refactoring. (see
92a6b16c9186 and about 200 previous changesets from Gregory Szorc)

While retesting, that is when '--retest' is enabled, only failure tests run
and others either skipped or ignored.

During retesting, "result.testsRun" holds the count of failure test that has
run. But as while printing output, we have subtracted the skipped and ignored
count from "result.testsRun". Therefore, to make the count remain
the same, we need to add skipped and ignored count before printing.